[feat] 과제 수정 및 삭제 기능 추가

+++ backend/pirocheck/src/main/java/backend/pirocheck/Assignment/repository/AssignmentItemRepository.java
@@ -1,15 +1,15 @@
 package backend.pirocheck.Assignment.repository;
 
-import backend.pirocheck.User.entity.User;
 import backend.pirocheck.Assignment.entity.AssignmentItem;
 import backend.pirocheck.Assignment.entity.AssignmentStatus;
+import backend.pirocheck.User.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface AssignmentItemRepository extends JpaRepository<AssignmentItem, Integer> {
+public interface AssignmentItemRepository extends JpaRepository<AssignmentItem, Long> {
 
     // 유저별 과제 목록 조회
     List<AssignmentItem> findByUserId(Long userId);
